Refactor squad name
